TST: Move test_associations_sdp_pools modules into jwst/regtest/associations_sdp_pools sub-directory for better organization.
