+++ jwst/regtest/associations_sdp_pools/test_associations_sdp_pools_slow_jw02739_1.py
@@ -2,13 +2,13 @@
 
 import pytest
 
-from jwst.regtest.assoc_rt_helpers import assoc_sdp_against_standard
+from jwst.regtest.associations_sdp_pools.assoc_rt_helpers import assoc_sdp_against_standard
 
 # Mark all tests in this module
 pytestmark = [pytest.mark.bigdata, pytest.mark.slow]
 
 
 def test_jw02739_20230710t150016_pool(_jail, rtdata, resource_tracker, request):
-    """This association tests the Asn_Lv3ImageMosaic rule."""
+    """Test the Asn_Lv3ImageMosaic rule."""
     pool_args = ("jw02739_20230710t150016_pool", ["-i", "c1000"])
     assoc_sdp_against_standard(rtdata, resource_tracker, request, pool_args)
