+++ jwst/regtest/associations_sdp_pools/test_associations_sdp_pools_slow_jw03702.py
@@ -8,12 +8,12 @@ Test using SDP-generated pools (slow).
 
 import pytest
 
-from jwst.regtest.assoc_rt_helpers import assoc_sdp_against_standard
+from jwst.regtest.associations_sdp_pools.assoc_rt_helpers import assoc_sdp_against_standard
 
 # Mark all tests in this module
 pytestmark = [pytest.mark.bigdata, pytest.mark.slow]
 
 
-def test_imprint_single_obs_1(_jail, rtdata, resource_tracker, request):
-    pool_args = ("jw01928_20250318t105523_pool", [])
+def test_imprint_2n_dithers_obs_1_2(_jail, rtdata, resource_tracker, request):
+    pool_args = ("jw03702_20250320t002824_pool", [])
     assoc_sdp_against_standard(rtdata, resource_tracker, request, pool_args)
