+++ jwst/regtest/associations_sdp_pools/test_associations_sdp_pools_slow_jw05221.py
@@ -8,12 +8,12 @@ Test using SDP-generated pools (slow).
 
 import pytest
 
-from jwst.regtest.assoc_rt_helpers import assoc_sdp_against_standard
+from jwst.regtest.associations_sdp_pools.assoc_rt_helpers import assoc_sdp_against_standard
 
 # Mark all tests in this module
 pytestmark = [pytest.mark.bigdata, pytest.mark.slow]
 
 
-def test_imprint_2n_dithers_obs_1_2(_jail, rtdata, resource_tracker, request):
-    pool_args = ("jw03702_20250320t002824_pool", [])
+def test_pool_032_nircam_wfss(_jail, rtdata, resource_tracker, request):
+    pool_args = ("jw05221_20250411t182730_pool", [])
     assoc_sdp_against_standard(rtdata, resource_tracker, request, pool_args)
