+++ jwst/regtest/associations_sdp_pools/test_associations_sdp_pools_slow_jw05308.py
@@ -8,12 +8,12 @@ Test using SDP-generated pools (slow).
 
 import pytest
 
-from jwst.regtest.assoc_rt_helpers import assoc_sdp_against_standard
+from jwst.regtest.associations_sdp_pools.assoc_rt_helpers import assoc_sdp_against_standard
 
 # Mark all tests in this module
 pytestmark = [pytest.mark.bigdata, pytest.mark.slow]
 
 
-def test_pool_024_nirspec_fss_nods_AND_SUBPXPTS4(_jail, rtdata, resource_tracker, request):
-    pool_args = ("jw03522_20250318t192624_pool", [])
+def test_pool_027_nirspec_ifu_nods_4NODS(_jail, rtdata, resource_tracker, request):
+    pool_args = ("jw05308_20250319t215735_pool", [])
     assoc_sdp_against_standard(rtdata, resource_tracker, request, pool_args)
