+++ jwst/regtest/associations_sdp_pools/test_associations_sdp_pools_slow_jw05398_2.py
@@ -8,16 +8,16 @@ Test using SDP-generated pools (slow).
 
 import pytest
 
-from jwst.regtest.assoc_rt_helpers import assoc_sdp_against_standard
+from jwst.regtest.associations_sdp_pools.assoc_rt_helpers import assoc_sdp_against_standard
 
 # Mark all tests in this module
 pytestmark = [pytest.mark.bigdata, pytest.mark.slow]
 
 
-def test_pool_034_wfss_parallel_NIRCAM_DEFAULT(_jail, rtdata, resource_tracker, request):
-    """This is "default" scenario: row grism, column grism, and direct image."""
+def test_pool_034_wfss_parallel_NIRCAM_3ROW_2DIRECT(_jail, rtdata, resource_tracker, request):
+    """Three row grism images and two direct images."""
     pool_args = (
-        "jw05398_20250312t000159_default_pool",
-        ["-i", "o041", "o042", "o043", "c1016"],
+        "jw05398_20250312t000159_3r2d_pool",
+        ["-i", "o030", "o031", "o032", "o033", "o035", "c1012", "c1013", "c1014"],
     )
     assoc_sdp_against_standard(rtdata, resource_tracker, request, pool_args)
